refactor app customizations and routes into separate files
